# docs/swagger.yaml
definitions:
  handlers.CreateAccountRequest:
    properties:
      document_number:
        example: "1234567890"
        type: string
    required:
    - document_number
    type: object
  handlers.CreateTransactionRequest:
    properties:
      account_id:
        example: 1
        type: integer
      amount:
        example: 123.45
        type: number
      operation_type_id:
        example: 4
        type: integer
    required:
    - account_id
    - amount
    - operation_type_id
    type: object
info:
  contact:
    email: sudip.post@gmail.com
    name: Sudip Bhandari
    url: https://sudipidus.github.io
  description: This is a simplified transaction service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pismo Transaction Service - Demo
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Greetings from Pismo-Test
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Greetings from Pismo-Test
      tags:
      - greeting/health-check
  /accounts:
    post:
      consumes:
      - application/json
      description: Create a new account
      parameters:
      - description: Create Account Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Create a new account
      tags:
      - accounts
  /accounts/{accountID}:
    get:
      consumes:
      - application/json
      description: Get an account by ID
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get an account by ID
      tags:
      - accounts
  /transactions:
    post:
      consumes:
      - application/json
      description: Create a new transaction
      parameters:
      - description: Create Transaction Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Create a new transaction
      tags:
      - transactions
swagger: "2.0"